backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.0
openai==1.12.0
websockets==12.0
pydantic==2.10.3
python-multipart==0.0.18
requests==2.31.0
amadeus